fix: separate Check componet
